fix: 사용자 검색 기능을 친구 API로 이관, 관련 테스트 코드 수정
