+++ OurMemory_Server/src/main/java/com/kds/ourmemory/controller/v1/notice/dto/NoticeDto.java
@@ -13,7 +13,7 @@ public class NoticeDto {
     @ApiModelProperty(value = "알림 번호")
     private final long noticeId;
 
-    @ApiModelProperty(value = "알림 종류", example = "FRIEND_REQUEST")
+    @ApiModelProperty(value = "알림 종류(FRIEND_REQUEST: 친구 요청)", example = "FRIEND_REQUEST")
     private final NoticeType type;
 
     @ApiModelProperty(value = "알림 문자열 값", example = "99")
@@ -22,7 +22,7 @@ public class NoticeDto {
     @ApiModelProperty(value = "알림 읽음 여부", example = "true")
     private final boolean read;
 
-    @ApiModelProperty(value = "알림 생성 날짜", notes = "yyyy-MM-dd HH:mm:ss")
+    @ApiModelProperty(value = "알림 생성 날짜(yyyy-MM-dd HH:mm:ss)", notes = "yyyy-MM-dd HH:mm:ss")
     private final String regDate;
 
     public NoticeDto(Notice notice) {
