+++ OurMemory_Server/src/main/java/com/kds/ourmemory/entity/notice/NoticeType.java
@@ -1,8 +1,10 @@
 package com.kds.ourmemory.entity.notice;
 
+import io.swagger.annotations.ApiModel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+@ApiModel
 @Getter
 @AllArgsConstructor
 public enum NoticeType {
